ai-projects/rag_service/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
transformers
torch
numpy
